==> notebooks/1_Combining_JUMP_Metadata.ipynb <==
**Notebook - Combining JUMP Metadata**

This notebook loads the different JUMP-DP dataset metadata files and combines them to form one dataframe containing all metadata information.

# Imports:

import pandas as pd
import numpy as np

# Load Metadata:
Files downloaded from: https://github.com/jump-cellpainting/datasets/tree/main/metadata

## Plates:

plates = pd.read_csv('../data/metadata/plate.csv.gz')
print(plates.shape)
plates.head(2)

## Wells:

wells = pd.read_csv('../data/metadata/well.csv.gz')
wells = wells.applymap(str)
print(wells.shape)
wells.head(2)

## Compounds:

compound = pd.read_csv('../data/metadata/compound.csv.gz')
print(compound.shape)
compound.head(2)

- 'JCP2022_999999' are untreated wells (these wells contain only cells) - see comment at https://github.com/jump-cellpainting/datasets/issues/49.

- This is altered below to change the n/a value for this row to 'NON_COMPOUND' instead so when the dataframes are merged there are no null values.

compound.loc[compound.Metadata_InChIKey.isnull(), ['Metadata_InChIKey', 'Metadata_InChI']] = 'NON_COMPOUND'

# Check that the above code ran successfully:
compound.loc[compound['Metadata_JCP2022'] == 'JCP2022_999999']

# Combining Metadata:

## Merging compound with well data:

c_well = pd.merge(compound, wells, on='Metadata_JCP2022', how='inner')
print(c_well.shape)
c_well.head()

- Non compound datapoints are retained, but flagged as they are without ChlKeys:

c_well[c_well['Metadata_JCP2022'] == 'JCP2022_999999'][0:2]

- Check for null values:

c_well[c_well.Metadata_Plate.isnull()]

## Merging compound + well data with plates data:

cwp_data = pd.merge(c_well, plates, on=['Metadata_Source', 'Metadata_Plate'] , how='inner')
print(cwp_data.shape)
cwp_data.head()

- Plate count check - for a specific batch, check the number of associated plates:

cwp_data.loc[cwp_data['Metadata_Batch'] == 'CP_28_all_Phenix1'].Metadata_Plate.value_counts()[0:5]

# Save Data:

cwp_data.to_csv('../data/cwp_data.csv', index=False)
