==> notebooks/2a_Preparing_MOA_Label_Data.ipynb <==
**Notebook - Preparing MOA Label Data**

This notebook loads compound mechanism of action information obtained from ChEMBL and the Drug Repurposing Hub and alters then to align with the information from the JUMP dataset metadata.

# Imports:

import pandas as pd
import numpy as np
import rdkit.Chem.inchi as inchi

# ChEMBL MOA Data:

See foot of https://github.com/jump-cellpainting/compound-annotator for data creation steps.

## ChEMBL annotation file

# Read in the ChEMBL annotation file:
chembl_df = pd.read_csv("../data/metadata/chembl/chembl_annotation.csv.gz")

# columns to drop from ChEMBL data:
drop_cols = ['assay_chembl_id','target_chembl_id','assay_type','pchembl_value',
             'confidence_score', 'pref_name']
chembl_df = chembl_df.drop(drop_cols, axis=1).drop_duplicates().reset_index(drop=True)
print(chembl_df.shape)
chembl_df.head(2)

## ChEMBL Kinase Mechanism Data:
Source: https://www.ebi.ac.uk/chembl/g/#browse/mechanisms_of_action

chembl_kinase = pd.read_csv('../data/metadata/chembl/chembl_kinase_moa.csv')
print(chembl_kinase.shape)
chembl_kinase.head(2)

## Combine MOA with inchi_key data:
'standard_inchi_key' in ChEMBL can be matched to 'Metadata_InChIKey' in the JUMP compound metadata, see here for more details: https://github.com/jump-cellpainting/compound-annotator

chembl_data = pd.merge(chembl_kinase, chembl_df, on='molecule_chembl_id', how='inner')
print(chembl_data.shape)
chembl_data.head(2)

# Rename columns to align with Drug Repo. Data:
chembl_data.rename(columns={"standard_inchi_key": "Metadata_InChIKey", "Smiles":"smiles",
                          "Target Name": "target", "Max Phase": "clinical_phase"}, inplace=True)
chembl_data['moa_src'] = 'chembl'
# Drop any duplicates with same ID and moa:
chembl_data = chembl_data.drop_duplicates(subset=['Metadata_InChIKey', 'moa'])

# Drug Repurposing Hub MOA Data:
Sourced from https://clue.io/repurposing#download-data
and https://github.com/jump-cellpainting/compound-annotator/blob/main/data/repurposing_samples_20200324_cleaned.txt.gz

## Drug Dataframe
Contains MOA labels.

drug_df = pd.read_csv('../data/metadata/clue_io/repurposing_drugs_20200324_cleaned.txt', 
                      encoding="ISO-8859-1", sep="\t", comment="!")

print(drug_df.shape)
drug_df.head(2)

## Sample Dataframe
Contains 'InChlKey', used to match to JUMP data.

sample_df = pd.read_csv('../data/metadata/clue_io/repurposing_samples_20200324_cleaned.txt', 
                        encoding="ISO-8859-1", sep="\t", comment="!")

print(sample_df.shape)
sample_df.head(2)

## Merge the Drug and Sample Dataframes:
See https://github.com/broadinstitute/lincs-cell-painting/tree/master/metadata/moa for previous usage of this approach.

drug_samp = drug_df.merge(sample_df, on="pert_iname", how="inner").reset_index(drop=True)

# Move broad_id to first column
col_order = drug_samp.columns.tolist()
col_order.insert(0, col_order.pop(col_order.index("broad_id")))
drug_samp = drug_samp.loc[:, col_order].assign(
    InChIKey14=drug_samp.InChIKey.apply(
        lambda x: inchi.InchiToInchiKey(x) if (x.startswith("InChI")) else x
    ).apply(lambda x: str(x)[:14])
)

# Output to file
output_file = "../data/metadata/clue_io/repurposing_moa"
drug_samp.rename(columns={"InChIKey": "Metadata_InChIKey"}, inplace=True)
# drug_samp.to_csv(f"{output_file}.csv", index=False)

print(drug_samp.shape)
drug_samp.head(2)

## Clean the DR Hub Data:

# Drop any rows where the compound key or the moa is Nan:
drug_samp = drug_samp.dropna(subset=['Metadata_InChIKey', 'moa'])

# Fill any Nan values in target column with 'Not listed'.:
drug_samp[['target']] = drug_samp[['target']].fillna(value='Not listed')

# Drop duplicates where Metadata_InChIKey and moa are the same (i.e. just a different vendor etc.)
drug_samp = drug_samp.drop_duplicates(subset=['Metadata_InChIKey', 'moa'])

# Subsetting the dataframe to only include inhibitors:
dhub_inhibitors = drug_samp[drug_samp['moa'].str.contains("inhibitor")].reset_index(drop=True)

# Assigning moa source:
dhub_inhibitors['moa_src'] = 'dr_hub'

# Manual changes to long-form MOA IDs:
Detailed below are the manual changes made to the above databases:

- Replaced "glycogen synthase kinase inhibitor" with GSK inhibitor.
- Replaced "Aurora kinase inhibitor" with AURK inhibitor.
- Replaced "rho associated kinase inhibitor" with ROCK inhibitor.
- Replaced "ALK tyrosine kinase receptor inhibitor" with ALK inhibitor.
- Replaced "Bruton's tyrosine kinase (BTK) inhibitor" with BTK inhibitor.
- Replaced "PDGFR tyrosine kinase receptor inhibitor" with PDGFR inhibitor.
- Altered any numbered CDK inhibitors i.e., "CDK9 inhibitor" to just "CDK inhibitor".
- Altered moas according to target column, i.e. EGFR|ERBB2 target = EGFR inhibitor|HER2 inhibitor
- Separated any BCR-ABL inhibitors into BCR and ABL by targets.

# Saving to Csv:

# Save to csv file:
chembl_data.to_csv('../data/metadata/chembl/chembl_cleaned.csv', index=False)

# Save to CSV file:
dhub_inhibitors.to_csv("../data/metadata/clue_io/dhub_inhibitors.csv", index=False)
