==> notebooks/2b_Aligning_MOA_to_Jump_Data.ipynb <==
**Notebook - Aligning MOA to Jump Data**

This notebook loads the JUMP-CP metadata and cross-references it with the Broad Drug Repurposing Hub and ChEMBL data to match compound chemical IDs ("InChIKey") in order to relate mechanisms of action to specific datapoints.

# Imports:

import pandas as pd
import numpy as np

# Load JUMP metadata:

cwp_data = pd.read_csv('../data/cwp_data.csv')
print(cwp_data.shape)
cwp_data.head(2)

# Load ChEMBL and Drug Repo. Hub (DRH) Data:

# Load ChEMBL Data:
chembl_moa = pd.read_csv("../data/metadata/chembl/chembl_cleaned.csv")
print(chembl_moa.shape)
chembl_moa.head(2)

# Load Drug Repo. Hub Data:
dhub_inhibitors = pd.read_csv("../data/metadata/clue_io/dhub_inhibitors.csv")
print(dhub_inhibitors.shape)
dhub_inhibitors.head(2)

# Merging DRH and ChEMBL MOA Data:

# Columns that both datasets share which will be used to merge:
keep_cols = ['moa', 'target', 'smiles', 'clinical_phase', 'Metadata_InChIKey', 'moa_src']
ds_reduced = dhub_inhibitors[keep_cols]
ch_reduced = chembl_moa[keep_cols]

# Concatenate the two dataframes together:
moa_combined = pd.concat([ds_reduced, ch_reduced]).reset_index(drop=True)
print(moa_combined.shape)
# Split out any datapoints with >1 moa:
moa_combined['moa'] = moa_combined['moa'].str.split('|')
moa_combined = moa_combined.explode('moa')
print(moa_combined.shape)
# Drop any duplicates with same ID and moa:
moa_combined = moa_combined.drop_duplicates(subset=['Metadata_InChIKey', 'moa'])
print(moa_combined.shape)
# Drop any rows with n/a values in:
moa_combined = moa_combined.dropna(subset=['Metadata_InChIKey', 'moa'])
print(moa_combined.shape)
moa_combined.head(2)

### Drop Columns Which Don't Contain Selected Kinase Inhibitors:

# Only retain the kinase inhibitors of interest:
ki_subs = moa_combined[moa_combined['moa'].str.contains(
    "PI3K|EGFR|HER2|p38 MAPK|JAK|RAF|AURK|ROCK|ALK|SRC|MEK|GSK|CDK|Bcr-Abl|PDGFR|FGFR|BTK|AKT|mTOR")
                      ].reset_index(drop=True)
print(ki_subs.shape)
ki_subs.head(3)

### Re-combining compounds with >1 MOA:

# Function to combine the moa's of datapoints with the same Metadata_InChIKey:
def combine_duplicates(group):
    # Find the unique values in each column
    moa = '|'.join(sorted(set('|'.join(group['moa']).split('|'))))
    target = '|'.join(sorted(set(group['target'].fillna('-'))))
    smiles = '|'.join(sorted(set(group['smiles'].fillna('-'))))
    clinical_phase = '|'.join(sorted(set(group['clinical_phase'].astype(str))))
    moa_src = '|'.join(sorted(set(group['moa_src'])))

    # Return a new dataframe with the combined data
    return pd.DataFrame({
        'moa': [moa],
        'target': [target],
        'smiles': [smiles],
        'clinical_phase': [clinical_phase],
        'Metadata_InChIKey': [group['Metadata_InChIKey'].iloc[0]],
        'moa_src': [moa_src],
    })

ki_comb = ki_subs.groupby('Metadata_InChIKey').apply(combine_duplicates).reset_index(drop=True) 
print(ki_comb.shape)
ki_comb.head(3)

### Checking duplicates have been merged correctly:

# Select one duplicate Metadata_InChIKey to inspect:
dupe_id = 'ZBNZXTGUTAYRHI-UHFFFAOYSA-N'

# Display it in original dataframe:
ki_subs[ki_subs['Metadata_InChIKey']==dupe_id]

# Check it's been combined correctly to form one datapoint with multiple moas:
ki_comb[ki_comb['Metadata_InChIKey']==dupe_id]

# Cross-referencing moa to metadata InChlKey:

### Align JUMP Data with MOAs from DRH/ Chembl:

cwp_ki = pd.merge(cwp_data, ki_comb, on='Metadata_InChIKey', how='inner')
print(cwp_ki.shape)
cwp_ki.head(3)

# Save/load Data:

cwp_ki.to_csv('../data/cwp_ki_moa.csv', index=False)
