==> notebooks/2c_U2OS_Kinase_Expression.ipynb <==
**Notebook - U2OS Kinase Expression**

This notebook looks at the U2OS cell line and associated kinase protein expression data. This information formed part of the basis for including/excluding specific kinase inhibitors.

Data Sources & Information:
- https://depmap.org/portal/download/all/?release=CCLE+2019&file=CCLE_RPPA_20181003.csv
- https://depmap.org/portal/cell_line/ACH-000364?tab=mutation
- https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7861052/
- https://gracebio.com/cancer-cell-line-encyclopedia-expanded-to-include-rppa-analysis-data/


# Imports:

import pandas as pd
import numpy as np

# Load Data:
Data Source: https://depmap.org/portal/download/all/?release=DepMap+Public+22Q4&file=OmicsExpressionProteinCodingGenesTPMLogp1.csv


"Gene expression TPM values of the protein coding genes for DepMap cell lines. Values are inferred from RNA-seq data using the RSEM tool and are reported after log2 transformation, using a pseudo-count of 1; log2(TPM+1)."

# Read in gene expression data:
u2os_omics = pd.read_csv('data/OmicsExpressionProteinCodingGenesTPMLogp1.csv')
u2os_omics = u2os_omics.rename(columns={u2os_omics.columns[0]: 'cell_type'})
print(u2os_omics.shape)
u2os_omics.head(2)

# U2OS Dep-Map ID - https://depmap.org/portal/cell_line/ACH-000364?tab=mutation:
u2os_depmap_id = 'ACH-000364'

# Subsetting the data to only include the U2OS cell line:
u2os_data = u2os_omics[u2os_omics['cell_type'].str.contains(u2os_depmap_id)]
u2os_data

# Example filtering:
u2os_data.filter(regex='EGFR')

# Relate DepMap Data to Kinase Inhibitor Classes:
Below are are the kinase inhibitor classes included within our dataset and their associated genes:

ki_tgt_dict = {
    'PI3K': 'PIK3CA|PIK3CB|PIK3CD|PIK3CG',
    'EGFR': 'EGFR',
    'p38 MAPK': 'MAPK11|MAPK12|MAPK13|MAPK14',
    'JAK': 'JAK1|JAK2|JAK3|TYK2',
    'RAF': '^ARAF|BRAF|^RAF1',
    'AURK': 'AURKA |AURKB|AURKC',
    'ALK': 'ALK ',
    'SRC': '^SRC |LYN |FYN|YES1|FGR|BLK|HCK|LCK',
    'ROCK': 'ROCK1|ROCK2',
    'MEK': 'MAP2K1|MAP2K2',
    'GSK': 'GSK3A|GSK3B',
    'CDK': 'CDK1 |CDK2 |CDK4 |CDK5 |CDK6 |CDK7 |CDK9 ',
    'VEGFR': 'FLT1|^KDR|FLT4',
    'BCR-ABL': '^ABL1 |^ABL2 |BCR',
    'PDGFR': 'PDGFRA|PDGFRB',
    'FGFR': 'FGFR1 |FGFR2|FGFR3|FGFR4',
    'BTK': '^BTK',    # '^' added to exclude IBTK
    'AKT': 'AKT1 |AKT2|AKT3',
    'mTOR': '^MTOR '
    }

u2os_data.filter(regex=ki_tgt_dict['EGFR'])

## Return Gene Expression Values:

exp_values = []

for ki in ki_tgt_dict:
    exp_data = u2os_data.filter(regex=ki_tgt_dict[ki]).T.reset_index().values
    exp_values.append(exp_data)

u2os_evs_df = pd.DataFrame(np.concatenate(exp_values), columns=['gene', 'exp_val'])
# Retain just gene names, removing numbers in brackets in original data:
u2os_evs_df['gene'] = [x[0] for x in u2os_evs_df.gene.str.split(" ", expand=False)]
# Reassigning gene to kinase inhibitor in dictionary:
ki_list = [[key for key, value in ki_tgt_dict.items() if x in value] for x in u2os_evs_df['gene']]
u2os_evs_df['k_inhib'] = [item for sublist in ki_list for item in sublist]
u2os_evs_df.head()

## Group by Kinase Inhibitor Class:

grouped_ki = pd.DataFrame(u2os_evs_df.groupby('k_inhib')['exp_val'].mean().reset_index()
                         ).sort_values('exp_val', ascending=False)
grouped_ki
