==> notebooks/5b_IBP_RF_SVM_XGBoost.ipynb <==
import pandas as pd
import numpy as np
import os
import pickle
from xgboost import XGBClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn import svm
from sklearn.model_selection import GridSearchCV, RandomizedSearchCV
from sklearn.metrics import accuracy_score, f1_score, roc_auc_score, precision_score, recall_score, average_precision_score
from modules.utils import load_cv
from modules.eval_funcs import cv_fold_eval
from modules.cleaning import rr_prefix

rand_seed = 42
moa_dict = {'PI3K' : 0, 'p38 MAPK': 1, 'RAF': 2, 'AURK': 3, 'CDK': 4, 'EGFR': 5, 'ROCK': 6,
             'MEK': 7, 'GSK': 8, 'mTOR': 9}

# Load IBP Data:

ki_ibp = pd.read_csv('data/ibp/ki_ibp.csv')
print(ki_ibp.shape)
ki_ibp.head(2)

# Load CV Fold Data:

# Load CV split data:
cv_splits = 5
cv_path = 'data/cv_splits/'
cv_data = load_cv(cv_path, cv_splits, ki_ibp, moa_dict, norm=None)

# Load Spherized and Harmonized Data:
pycy_spher = 'output/MAD_Sphere/MADS_PyCyFS.pkl'
shap_spher = 'output/MAD_Sphere/MADS_ShapFS.pkl'
pycy_harm = 'output/MAD_Harmony/MAD_Harmony_PyCyFS.pkl'
shap_harm = 'output/MAD_Harmony/MAD_Harmony_ShapFS.pkl'

with open(shap_spher, 'rb') as file:
    loaded_data = pickle.load(file)
    
# Specify the prefix to remove
prefix_to_remove = 'sph_'
loaded_data = rr_prefix(loaded_data, prefix_to_remove)

shap_spher = 'output/Corrected/MADS_ShapFS.pkl'

with open(shap_spher, 'rb') as file:
    loaded_data = pickle.load(file)

    # Specify the prefix to remove
prefix_to_remove = 'sph_'
loaded_data = rr_prefix(loaded_data, prefix_to_remove)

loaded_data[0].keys()

# Load Feature Selection Data:

cv_results = pd.read_csv('output/feat_sel/shap_fs_res.csv')

# Random Forest:

## Baseline Model:

rf_clf = RandomForestClassifier(random_state=rand_seed)

### Minmax Norm, no Feature Selection:

cv_fold_eval(rf_clf, cv_data)

### MADS + Shapley FS:

cv_fold_eval(rf_clf, loaded_data)

### MADS + Pycy FS:

cv_fold_eval(rf_clf, loaded_data)

## MADH + Shap FS:

cv_fold_eval(rf_clf, loaded_data)

### MADH + PyCy FS

cv_fold_eval(rf_clf, loaded_data)

## Random Search:

# Number of trees in random forest
n_estimators = [int(x) for x in np.linspace(start = 16, stop = 1024, num = 6)]
# Maximum number of levels in tree
max_depth = [int(x) for x in np.linspace(2, 64, num = 8)]
# Minimum number of samples required to split a node
min_samples_split = [2, 6, 8, 12]
# Minimum number of samples required at each leaf node
min_samples_leaf = [1, 2, 4]
# Method of selecting samples for training each tree
bootstrap = [True, False]

# Create the random grid
random_grid = {'n_estimators': n_estimators,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap}
print(random_grid)

for cv in range(cv_splits):
    # Define model and search:
    rand_search_rf = RandomForestClassifier(random_state = rand_seed) 
    
    # Search across 100 different combinations, and use all available cores
    rf_random = RandomizedSearchCV(estimator = rand_search_rf, param_distributions = random_grid, 
                               n_iter = 25, verbose=1, random_state=rand_seed, n_jobs = -1)
    
    # Fit Random Search Model:
    rf_random.fit(loaded_data[cv]['X_train'], loaded_data[cv]['y_train'])
    
    print("Best CV Accuracy: ", rf_random.best_score_)
    print("Best Parameters: ", rf_random.best_params_)
    print("--------------------------------------------")

### Re-Fit Model with Hyp. Tuned Params:

hyp_rf_clf = RandomForestClassifier(n_estimators=474, min_samples_split=2,
                                    min_samples_leaf=1, max_depth=55, bootstrap=False,
                                    random_state=rand_seed)

### Minmax + No FS:

cv_fold_eval(hyp_rf_clf, cv_data)

### MADS + Shap FS:

hyp_rf_clf = RandomForestClassifier(n_estimators=822, min_samples_split=2,
                                    min_samples_leaf=1, max_depth=28, bootstrap=False,
                                    random_state=rand_seed)

cv_fold_eval(hyp_rf_clf, loaded_data)

## Apply Feature Selection:

n_feats = 150
top_feats = cv_results['features'][0:n_feats].tolist()
cv_fold_eval(hyp_rf_clf, cv_data, top_feats)

# XGBoost:

## Baseline:

xgb_model = XGBClassifier()

### Minmax and no FS:

cv_fold_eval(xgb_model, cv_data)

## Random Search:

rs_params = {'max_depth': [3, 5, 6, 10, 15, 20],
           'learning_rate': [0.01, 0.1, 0.2, 0.3],
           'subsample': np.arange(0.5, 1.0, 0.1),
           'colsample_bytree': np.arange(0.4, 1.0, 0.1),
           'colsample_bylevel': np.arange(0.4, 1.0, 0.1),
           'n_estimators': np.arange(50, 1000, 200)}

rs_params

for cv in range(cv_splits):
    # Define model and search:
    rand_search_XGB = XGBClassifier(random_state=rand_seed) 
    
    # Search across 100 different combinations, and use all available cores
    xgb_random = RandomizedSearchCV(estimator=rand_search_XGB, param_distributions=rs_params, 
                               n_iter=50, verbose=1, random_state=rand_seed)
    
    # Fit Random Search Model:
    xgb_random.fit(loaded_data[cv]['X_train'], loaded_data[cv]['y_train'])
    
    print("Best CV Accuracy: ", xgb_random.best_score_)
    print("Best Parameters: ", xgb_random.best_params_)
    print("--------------------------------------------")

### Re-Fit Model with Hyp. Tuned Params:

xgb_model = XGBClassifier(n_estimators = 400, subsample=0.5, colsample_bytree=0.8, learning_rate = 0.05, 
                          max_depth=15, colsample_bylevel=0.4)

cv_fold_eval(xgb_model, cv_data)

### MADS + Shap:

fold_dict = cv_fold_eval(xgb_model, loaded_data)

- 96 compounds:

fold_dict = cv_fold_eval(xgb_model, loaded_data)

- Above shows crude statistics calculated as mean values across folds, below are the precise metrics, calculated when all predictions across test sets are combined before metrics are calculated:

# Concatenate the arrays within the results lists:
pred_arr = fold_dict['pred_arr']
act_arr = fold_dict['act_arr']
proba_arr = np.vstack(fold_dict['proba_arr'])

# Calculate performance metrics at a compound-level:
acc = accuracy_score(act_arr, pred_arr)
f1 = f1_score(act_arr, pred_arr, average='macro')
precision = precision_score(act_arr, pred_arr, average='macro', zero_division=0)
recall = recall_score(act_arr, pred_arr, average='macro')
roc_auc = roc_auc_score(act_arr, proba_arr, average='macro', multi_class='ovr')

# Calculate AUPR for each class
aupr_scores = [average_precision_score(act_arr == class_index, proba_arr[:, class_index]
                                       ) for class_index in range(proba_arr.shape[1])]
mean_aupr = np.mean(aupr_scores)

# Print metrics:
print('Accuracy: {:.2f}%'.format(acc*100))
print('F1 Score: {:.2f}%'.format(f1*100))
print('Precision: {:.2f}%'.format(precision * 100))
print('Recall: {:.2f}%'.format(recall * 100))
print('ROC AUC: {:.2f}%'.format(roc_auc * 100))
print('AUPR: {:.2f}%'.format(mean_aupr * 100))

### MADS + Pycy:

cv_fold_eval(xgb_model, loaded_data)

### MADH + Shap:

xgb_model = XGBClassifier(n_estimators = 650, subsample=0.7, colsample_bytree=0.8, 
                          learning_rate = 0.1, max_depth=15, colsample_bylevel=0.7)

cv_fold_eval(xgb_model, loaded_data)

### MADH + Pycy:

cv_fold_eval(xgb_model, loaded_data)

# SVM:

## Baseline:

svm_clf = svm.SVC(random_state = rand_seed, probability=True)

cv_fold_eval(svm_clf, cv_data)

## Random Search:

rs_params = {'C': [0.1, 0.2, 0.5, 1, 10, 100, 1000], 
              'gamma': [1, 0.1, 0.01, 0.001, 0.0001],
              'kernel': ['poly', 'rbf', 'linear']} 

for cv in range(cv_splits):
    # Define model and search:
    svm_clf = svm.SVC(random_state = rand_seed, probability=True) 
    
    # Search across 100 different combinations, and use all available cores
    svm_random = RandomizedSearchCV(estimator=svm_clf, param_distributions=rs_params, 
                               n_iter=50, verbose=1, random_state=rand_seed)
    
    # Fit Random Search Model:
    svm_random.fit(loaded_data[cv]['X_train'], loaded_data[cv]['y_train'])
    
    print("Best CV Accuracy: ", svm_random.best_score_)
    print("Best Parameters: ", svm_random.best_params_)
    print("--------------------------------------------")

svm_clf = svm.SVC(kernel = 'rbf', gamma = 0.001, C = 100, 
                  random_state = rand_seed, probability=True)
cv_fold_eval(svm_clf, loaded_data)

# Training Full XGBoost Model:

# Load full spherized data with Shapley FS:
with open('output/MAD_Sphere/MADS_FULL_ShapFS.pkl', 'rb') as file:
    full_sph_data = pickle.load(file)

xgb_model = XGBClassifier(n_estimators = 400, subsample=0.5, colsample_bytree=0.8, learning_rate = 0.05, 
                          max_depth=15, colsample_bylevel=0.4)

# Fit Model on full data:
xgb_model.fit(full_sph_data['sph_X_data'], full_sph_data['sph_y_data'])

with open("output/XGB/xgb_full_model.pkl", "wb") as model_file:
    pickle.dump(xgb_model, model_file)
