==> notebooks/7_Model_Results.ipynb <==
**Notebook - Model Results**

This notebook combines the cross-validated results from each cv-fold that the model was evaluated on and combines the predictions before calculating performance metrics across all test cases.

import pandas as pd
import glob
import re
import matplotlib.pyplot as plt
import matplotlib as mpl
import numpy as np
from sklearn.metrics import (accuracy_score, f1_score, roc_auc_score, precision_score, auc,
                             recall_score, average_precision_score)
from modules.viz import conf_mat
from modules.eval_funcs import amalgamate_preds

cv_folds = 5

def string_to_float_list(s, method='new'):
    if method == 'old':
        s = s.replace('[', '').replace('[', '').replace('\n', '')
        s = re.sub(r'\s+', ', ', re.sub(r'\s*\]', '', s))
#         print("STRING", s)
    return [float(val) for val in s.split(',')]

# Load Data:

def summary_stats(df):
    """
    Calculate summary statistics using combined CV results.
    """
    if 'actual' in df.columns:
        df.rename(columns={'actual': 'actuals'}, inplace=True)
        
    acc = accuracy_score(df.actuals, df.preds)
    f1 = f1_score(df.actuals, df.preds, average='macro')
    precision = precision_score(df.actuals, df.preds, average='macro', zero_division=0)
    recall = recall_score(df.actuals, df.preds, average='macro')
    
    # Create a prediction array from probs:
    pred_arr = np.array(df['proba'].tolist())
    roc_auc = roc_auc_score(df.actuals, pred_arr, average='macro', multi_class='ovr')
    
    # Calculate AUPR for each class
    aupr_scores = [average_precision_score(df.actuals == class_index, pred_arr[:, class_index]
                                          ) for class_index in range(pred_arr.shape[1])]
    mean_aupr = np.mean(aupr_scores)
    
    return {'acc': acc, 'f1': f1, 'precision': precision, 'recall': recall,
           'roc_auc': roc_auc, 'aupr': mean_aupr}

# Locations of model results:
swin_img = 'results/swinv2/model_eval/'  # SwinV2 model results
cvf_fusion = 'results/fusion/model_eval/'  # CVF model results
ibp_mlp = 'results/IBP_MLP/model_eval'  # IBP MLP model results

# Load Kinase Inhibitor metadata:
ki_meta = pd.read_csv('data/cwp_ki_moa.csv')
print(ki_meta.shape)
ki_meta.head(2)

## Return Prediction Dataframes:

def return_preds(pred_dir, method, cv_folds=5):
    """
    Returns the five prediction dataframes associated with the cross-val test datasets for the specified model.
    method: the method of storing proba data changed, so this reflects how it was stored when each model was run, don't change below
    """
    cpnd_frames = []
    well_frames = []

    for i in range(cv_folds):
        cpnd_file = glob.glob(f'{pred_dir}/CV{i}_eval/cpnd*.csv')
        well_file = glob.glob(f'{pred_dir}/CV{i}_eval/well*.csv')
        cdf = pd.read_csv(cpnd_file[0])
        wdf = pd.read_csv(well_file[0])
        cpnd_frames.append(cdf)
        well_frames.append(wdf)

    # Create combined dataframes of predictions across folds:
    cpnd_comb = pd.concat(cpnd_frames, axis=0).reset_index(drop=True)
    well_comb = pd.concat(well_frames, axis=0).reset_index(drop=True)
    cpnd_comb['proba'] = cpnd_comb['proba'].apply(string_to_float_list, method=method)
    well_comb['proba'] = well_comb['proba'].apply(string_to_float_list, method=method)
    
    return cpnd_comb, well_comb

# Summary Statistics:

## Swin Image:

swin_cpnd, swin_well = return_preds(swin_img, method='new')

- Performance metrics at a compound level:

swin_cpnd_df = amalgamate_preds(swin_cpnd, 'cpnd_ids')
summary_stats(swin_cpnd_df)

- Performance metrics at a well level:

summary_stats(swin_well)

## IBP MLP:

ibp_cpnd, ibp_well = return_preds(ibp_mlp, method='old')

summary_stats(ibp_cpnd)

summary_stats(ibp_well)

## Cell-Vision Fusion:

fuse_cpnd, fuse_well = return_preds(cvf_fusion, method='old')

fuse_cpnd_df = amalgamate_preds(fuse_cpnd, 'cpnd_ids')
summary_stats(fuse_cpnd_df)

summary_stats(fuse_well)
